+++ src/core/server/events/playerDeath.ts
@@ -3,10 +3,24 @@ import { playerFuncs } from '../extensions/Player';
 
 alt.on('playerDeath', handleDeath);
 
-function handleDeath(player: alt.Player, killer: alt.Player, weaponHash: any): void {
-    if (!player || !player.valid) {
-        return;
+function handleDeath(player: alt.Player, killer: alt.Entity, weaponHash: any): void {
+    if (player && player.valid) {
+        playerFuncs.set.dead(player, weaponHash);
     }
 
-    playerFuncs.set.dead(player, killer, weaponHash);
+    if (killer instanceof alt.Player && player !== killer) {
+        alt.log(`(${player.id}) ${player.data.name} killed by ${killer.data.name}.`);
+    }
+
+    if (killer instanceof alt.Player && player === killer) {
+        alt.log(`(${player.id}) ${player.data.name} died.`);
+    }
+
+    if (killer instanceof alt.Vehicle && killer.driver) {
+        alt.log(`(${player.id}) ${player.data.name} was killed in vehicular combat by ${killer.driver.data.name}`);
+    }
+
+    if (killer instanceof alt.Vehicle && !killer.driver) {
+        alt.log(`(${player.id}) ${player.data.name} killed themself in a vehicular incident.`);
+    }
 }